[Feature] Company bizNo(사업자 등록번호) 타입 변경

+++ module-core/src/main/java/com/back2basics/company/port/in/command/CreateCompanyCommand.java
@@ -10,7 +10,7 @@ public class CreateCompanyCommand {
     private String name;
     private String ceoName;
     private String bio;
-    private Integer bizNo;
+    private Long bizNo;
     private String address;
     private String email;
     private String tel;
